Bump NumPy lower bound to 1.16
In preparation for building this feedstock for aarch64 and ppc64le, bump
the `numpy` minimum version to 1.16, which is the version we have built
on those architectures. It was found to be too difficult to build
earlier versions on those architectures; so, the minimum `numpy` version
was bumped to the oldest version that could reasonably be built.